api/requirements.txt
Flask==0.12.0
PyYAML==3.12.0
